# src/data/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "0.0.1",
    "title": "JaRcom PRO API 1.0",
    "description": "JaRcom PRO back-end API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://backend.jarcom.pro"
    },
    {
      "url": "http://backend.jarcom.pro"
    }
  ],
...
  "paths": {
...
    "/news": {
      "get": {
        "tags": [
          "Other"
        ],
        "summary": "Get all news",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search by title",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "OK \"Successful operation\" (array of objects)",
            "content": {
              "application/json": {
                "schema": {
                  "title": "200",
                  "type": "object",
                  "properties": {
                    "_id": {
                      "example": "63f8635ce9466b36765fbf24"
                    },
                    "title": {
                      "example": "text title"
                    },
                    "url": {
                      "example": "url to news"
                    },
                    "description": {
                      "example": "text body"
                    },
                    "date": {
                      "example": "12.08.2022"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/userAndPets": {
      "get": {
        "tags": [
          "Other"
        ],
        "summary": "Get personal information about the user, and information about the user's pets",
        "parameters": [],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK \"Successful operation\" (object and array of objects)",
            "content": {
              "application/json": {
                "schema": {
                  "title": "200",
                  "type": "object",
                  "properties": {
                    "user": {
                      "example": {
                        "email": "user email",
                        "cityRegion": "address",
                        "mobilePhone": "phone number",
                        "birthdate": "birthdate",
                        "photoURL": "url to image",
                        "name": "user name"
                      }
                    },
                    "pets": {
                      "example": [
                        {
                          "_id": "petId",
                          "date": "birthdate",
                          "breed": "breed",
                          "comments": "comments",
                          "photoURL": "url to image",
                          "name": "pet name"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/uploadAvatar": {
      "post": {
        "tags": [
          "Other"
        ],
        "summary": "Upload a picture/photo/avatar and get a link to it",
        "parameters": [],
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK \"Successful operation\"",
            "content": {
              "application/json": {
                "schema": {
                  "title": "200",
                  "type": "object",
                  "properties": {
                    "urlAvatar": {
                      "type": "string",
                      "description": "description",
                      "example": "url"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}